+++ js/content.js
@@ -187,7 +187,7 @@ chrome.runtime.onMessage.addListener(function(msg, sender, sendResponse) {
                 case 'printLabelReclaim':
                 case 'printLabelRepair':
                 case 'printLabelRestock':
-                    if (ticketType !== 'task') {
+                    if (msg.type !== 'printLabelBroken' && ticketType !== 'task') {
                         sendResponse({success: false, errMsg: 'Unable to detect an open task.'});
                     } else {
                         injectData = {type: msg.type}